requirements.txt
numpy
matplotlib
h5py
scipy
Pillow
scipy
scikit-image
